# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AoraGrand Hotel Booking API",
        "description": "RESTful API for AoraGrand Hotel Booking System with Midtrans payment integration",
        "contact": {
            "name": "AoraGrand Hotel Support",
            "email": "admin@aoragrand.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://api.aoragrand.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get admin dashboard overview",
                "operationId": "f6e3ffeb3c9fb988ca2a77ad48e2c04c",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard data retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "overview": {
                                                    "properties": {
                                                        "total_users": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "total_bookings": {
                                                            "type": "integer",
                                                            "example": 75
                                                        },
                                                        "total_revenue": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 50000000
                                                        },
                                                        "occupancy_rate": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 85.5
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "recent_bookings": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Booking"
                                                    }
                                                },
                                                "room_status": {
                                                    "properties": {
                                                        "available": {
                                                            "type": "integer",
                                                            "example": 45
                                                        },
                                                        "occupied": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "maintenance": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "cleaning": {
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/dashboard/stats": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get detailed statistics",
                "operationId": "7049f743c35dc1ca33c9b828cb34dee9",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Statistics period",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "daily",
                                "weekly",
                                "monthly"
                            ],
                            "example": "monthly"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statistics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "booking_trends": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "revenue_trends": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "room_type_popularity": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "payment_methods": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {
                "title": "Booking",
                "description": "Booking model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "booking_code": {
                        "type": "string",
                        "example": "BK20231201ABC123"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "room_type": {
                        "$ref": "#/components/schemas/RoomType"
                    },
                    "room": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "room_number": {
                                "type": "string",
                                "example": "STD101"
                            },
                            "status": {
                                "type": "string",
                                "example": "available"
                            }
                        },
                        "type": "object"
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-01"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-03"
                    },
                    "nights": {
                        "type": "integer",
                        "example": 2
                    },
                    "guests": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1000000
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "checked_in",
                            "checked_out",
                            "cancelled"
                        ],
                        "example": "confirmed"
                    },
                    "special_requests": {
                        "type": "string",
                        "example": "Late check-in"
                    },
                    "payment": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "payment_code": {
                                "type": "string",
                                "example": "PAY20231201XYZ789"
                            },
                            "status": {
                                "type": "string",
                                "example": "paid"
                            },
                            "amount": {
                                "type": "number",
                                "format": "float",
                                "example": 1000000
                            }
                        },
                        "type": "object"
                    },
                    "confirmed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "RoomType": {
                "title": "Room Type",
                "description": "Room Type model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Standard Room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comfortable standard room with basic amenities"
                    },
                    "price_per_night": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_rooms": {
                        "type": "integer",
                        "example": 20
                    },
                    "available_rooms": {
                        "type": "integer",
                        "example": 15
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "AC",
                            "TV",
                            "WiFi"
                        ]
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "/images/room1.jpg",
                            "/images/room2.jpg"
                        ]
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+62812345678"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Contoh No. 123"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ],
                        "example": "male"
                    },
                    "identity_number": {
                        "type": "string",
                        "example": "1234567890123456"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user"
                        ],
                        "example": "user"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}